refactor: separated test by method
