+++ src/main/resources/db/changelog/0.8.0/project_steps.csv
@@ -1,7 +1,7 @@
 id,name,project_id,priority_id,assigned_employee_id,deadline,version,created,last_modified,is_deleted
 0565848e-6138-413c-a80b-e53153f30f89,step 1,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-08-25 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,FALSE
-fb91749e-5a5f-4953-bb85-b176d98fb2fa,step 1.1 to delete,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-08-25 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,FALSE
+fb91749e-5a5f-4953-bb85-b176d98fb2fa,step 1.1 to delete,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-08-26 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,FALSE
 7f1ca6e5-25e8-4bee-9829-fa70cf672895,step 2,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-09-25 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,FALSE
-fda8d9fd-fe63-4b89-9c9a-96cd33e616ff,step 2.2 deleted,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-09-25 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,TRUE
+fda8d9fd-fe63-4b89-9c9a-96cd33e616ff,step 2.2 deleted,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-09-26 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,TRUE
 39b12d4c-f58c-473a-b9cf-59bd392e1aef,step 3,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-10-25 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,FALSE
-e6c8acd6-a7bb-48a6-bad7-e4154b5010f0,step to update,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-10-25 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,FALSE
+e6c8acd6-a7bb-48a6-bad7-e4154b5010f0,step to update,6c89f693-bb0b-41c0-aebd-4abac36b49d3,da12b38a-ce6d-4548-a99e-01a4b64e1244,92a6901f-9131-4cdd-a363-a6cefea8f865,2025-10-26 00:00:00.000,0,2025-07-25 00:00:00.000,2025-07-25 00:00:00.000,FALSE
