+++ src/test/java/com/pawelnu/projectmanager/endpoints/project/step/ProjectStepControllerGetByIdTest.java
@@ -1,0 +1,146 @@
+package com.pawelnu.projectmanager.endpoints.project.step;
+
+import static com.pawelnu.projectmanager.utils.Utils.invalidUUIDError;
+import static com.pawelnu.projectmanager.utils.Utils.unauthorizedError;
+import static com.pawelnu.projectmanager.utils.Utils.withBadJwt;
+import static com.pawelnu.projectmanager.utils.Utils.withJwt;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.pawelnu.projectmanager.config.security.jwt.JwtUtils;
+import com.pawelnu.projectmanager.endpoints.project.step.dto.ProjectStepDTO;
+import com.pawelnu.projectmanager.exception.NotFoundException;
+import com.pawelnu.projectmanager.exception.model.ReactAdminError;
+import com.pawelnu.projectmanager.utils.Consts.MSG;
+import com.pawelnu.projectmanager.utils.Path;
+import com.pawelnu.projectmanager.utils.Utils;
+import com.pawelnu.projectmanager.utils.Utils.Postgres;
+import com.pawelnu.projectmanager.utils.Utils.SpringDataSource;
+import java.time.Instant;
+import java.util.UUID;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpStatus;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
+@Testcontainers
+@ActiveProfiles("test")
+@Slf4j
+class ProjectStepControllerGetByIdTest {
+  @Autowired private JwtUtils jwtUtils;
+  @Autowired private MockMvc mockMvc;
+  @Autowired private ObjectMapper objectMapper;
+  private static final String BASE_URL = "/" + Path.API_PROJECT_STEPS;
+
+  @Container
+  static PostgreSQLContainer<?> postgres =
+      new PostgreSQLContainer<>(Postgres.POSTGRES_17)
+          .withDatabaseName(Postgres.DB_NAME)
+          .withUsername(Postgres.USER)
+          .withPassword(Postgres.PASSWORD);
+
+  @DynamicPropertySource
+  static void postgresProperties(DynamicPropertyRegistry registry) {
+    registry.add(SpringDataSource.URL, postgres::getJdbcUrl);
+    registry.add(SpringDataSource.USERNAME, postgres::getUsername);
+    registry.add(SpringDataSource.PASSWORD, postgres::getPassword);
+  }
+
+  @BeforeEach
+  void beforeEach() {
+    Utils.generateToken(jwtUtils);
+  }
+
+  @Test
+  void shouldReturn_200_getProjectStepById() throws Exception {
+    String projectStepId = "0565848e-6138-413c-a80b-e53153f30f89";
+    String url = BASE_URL + "/" + projectStepId;
+    MvcResult response = mockMvc.perform(get(url).with(withJwt())).andReturn();
+    int status = response.getResponse().getStatus();
+    String contentAsString = response.getResponse().getContentAsString();
+    ProjectStepDTO responseBody = objectMapper.readValue(contentAsString, ProjectStepDTO.class);
+    assertEquals(HttpStatus.OK.value(), status);
+    assertEquals(UUID.fromString(projectStepId), responseBody.getId());
+    assertEquals("step 1", responseBody.getName());
+    assertEquals("steps checker", responseBody.getProjectName());
+    assertEquals("VERY_LOW", responseBody.getPriorityValue());
+    assertEquals("Alleen Koepp", responseBody.getAssignedEmployee());
+    assertEquals(Instant.parse("2025-08-25T00:00:00Z"), responseBody.getDeadline());
+  }
+
+  @Test
+  void shouldReturn_400_getProjectStepById() throws Exception {
+    String projectStepId = "invalid-uuid";
+    String url = BASE_URL + "/" + projectStepId;
+    MvcResult response = mockMvc.perform(get(url).with(withJwt())).andReturn();
+    int status = response.getResponse().getStatus();
+    String contentAsString = response.getResponse().getContentAsString();
+    ReactAdminError responseBody = objectMapper.readValue(contentAsString, ReactAdminError.class);
+    assertEquals(HttpStatus.BAD_REQUEST.value(), status);
+    assertEquals(invalidUUIDError(), responseBody);
+  }
+
+  @Test
+  void shouldReturn_401_getProjectStepById() throws Exception {
+    String projectStepId = "cf578fec-006b-4604-a5e8-5ad1b3ea2be5";
+    String url = BASE_URL + "/" + projectStepId;
+    MvcResult response = mockMvc.perform(get(url)).andReturn();
+    int status = response.getResponse().getStatus();
+    String contentAsString = response.getResponse().getContentAsString();
+    ReactAdminError responseBody = objectMapper.readValue(contentAsString, ReactAdminError.class);
+    assertEquals(HttpStatus.UNAUTHORIZED.value(), status);
+    assertEquals(unauthorizedError(), responseBody);
+  }
+
+  @Test
+  void shouldReturn_403_getProjectStepById() throws Exception {
+    String projectStepId = "cf578fec-006b-4604-a5e8-5ad1b3ea2be5";
+    String url = BASE_URL + "/" + projectStepId;
+    MvcResult response = mockMvc.perform(get(url).with(withBadJwt())).andReturn();
+    int status = response.getResponse().getStatus();
+    String contentAsString = response.getResponse().getContentAsString();
+    ReactAdminError responseBody = objectMapper.readValue(contentAsString, ReactAdminError.class);
+    assertEquals(HttpStatus.FORBIDDEN.value(), status);
+    assertEquals(Utils.accessDeniedError(), responseBody);
+  }
+
+  @Test
+  void shouldReturn_404_getProjectStepById() throws Exception {
+    String projectStepId = "89258385-aa00-47d3-bafe-88bc1d56e6ee";
+    String url = BASE_URL + "/" + projectStepId;
+    MvcResult response = mockMvc.perform(get(url).with(withJwt())).andReturn();
+    int status = response.getResponse().getStatus();
+    String contentAsString = response.getResponse().getContentAsString();
+    NotFoundException responseBody =
+        objectMapper.readValue(contentAsString, NotFoundException.class);
+    assertEquals(HttpStatus.NOT_FOUND.value(), status);
+    assertEquals(MSG.PROJECT_STEP_NOT_FOUND + projectStepId, responseBody.getMessage());
+  }
+
+  @Test
+  void shouldReturn_404_getProjectStepById_isDeletedTrue() throws Exception {
+    String projectStepId = "9d93bd6d-e48a-4bad-b4f5-6f71d27c0342";
+    String url = BASE_URL + "/" + projectStepId;
+    MvcResult response = mockMvc.perform(get(url).with(withJwt())).andReturn();
+    int status = response.getResponse().getStatus();
+    String contentAsString = response.getResponse().getContentAsString();
+    NotFoundException responseBody =
+        objectMapper.readValue(contentAsString, NotFoundException.class);
+    assertEquals(HttpStatus.NOT_FOUND.value(), status);
+    assertEquals(MSG.PROJECT_STEP_NOT_FOUND + projectStepId, responseBody.getMessage());
+  }
+}
